Split post-render stuff into separate file
